refactor(user-management): relocate audit.py to autobot_shared, update mover changelog (#12647)
