+++ autobot_shared/user_management/models/audit.py
@@ -1,0 +1,205 @@
+# Copyright 2025-2026 mrveiss
+# SPDX-License-Identifier: Apache-2.0
+# AutoBot - AI-Powered Automation Platform
+# Author: mrveiss
+"""
+Audit Log Model (#12647)
+
+Tracks all security-relevant actions for compliance.
+
+Not byte-identical between ``autobot-backend`` and ``autobot-slm-backend``
+before this move, but the diff was comment-wording only, no schema, value, or
+behavior difference: the backend's ``# nosec``/``# nosemgrep`` suppression
+comments were more verbose, and its ``PERMISSION_DENIED`` constant carried a
+cross-reference note tying its string value to SLM's copy — a note that
+stops being meaningful once both files are one canonical module. This file
+keeps SLM's shorter comment wording; every enum/constant value is unchanged.
+"""
+
+import uuid
+from datetime import datetime
+
+from sqlalchemy import DateTime, ForeignKey, Index, String, func
+from sqlalchemy.dialects.postgresql import JSONB, UUID
+from sqlalchemy.orm import Mapped, mapped_column
+
+from autobot_shared.user_management.models.base import Base
+
+
+class AuditLog(Base):
+    """
+    Audit log entry.
+
+    Records all security-relevant actions for compliance and debugging.
+    Immutable - entries are never updated or deleted.
+    """
+
+    __tablename__ = "audit_logs"
+
+    id: Mapped[uuid.UUID] = mapped_column(
+        UUID(as_uuid=True),
+        primary_key=True,
+        default=uuid.uuid4,
+    )
+
+    # Organization context (nullable for platform-level events)
+    org_id: Mapped[uuid.UUID | None] = mapped_column(
+        UUID(as_uuid=True),
+        ForeignKey("organizations.id", ondelete="SET NULL"),
+        nullable=True,
+        index=True,
+    )
+
+    # Actor (who performed the action)
+    user_id: Mapped[uuid.UUID | None] = mapped_column(
+        UUID(as_uuid=True),
+        ForeignKey("users.id", ondelete="SET NULL"),
+        nullable=True,
+        index=True,
+    )
+
+    # Action performed
+    action: Mapped[str] = mapped_column(
+        String(100),
+        nullable=False,
+        index=True,
+    )
+
+    # Resource type affected
+    resource_type: Mapped[str] = mapped_column(
+        String(100),
+        nullable=False,
+        index=True,
+    )
+
+    # Resource ID affected
+    resource_id: Mapped[uuid.UUID | None] = mapped_column(
+        UUID(as_uuid=True),
+        nullable=True,
+    )
+
+    # Outcome
+    outcome: Mapped[str] = mapped_column(
+        String(50),
+        default="success",
+        nullable=False,
+    )
+
+    # Additional details (JSON)
+    details: Mapped[dict] = mapped_column(
+        JSONB,
+        default=dict,
+        nullable=False,
+    )
+
+    # Request context
+    ip_address: Mapped[str | None] = mapped_column(
+        String(45),  # IPv6 max length
+        nullable=True,
+    )
+
+    user_agent: Mapped[str | None] = mapped_column(
+        String(500),
+        nullable=True,
+    )
+
+    # Timestamp (immutable)
+    created_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True),
+        server_default=func.now(),
+        nullable=False,
+        index=True,
+    )
+
+    # Indexes for common queries
+    __table_args__ = (
+        Index("idx_audit_org_created", "org_id", "created_at"),
+        Index("idx_audit_user_created", "user_id", "created_at"),
+        Index("idx_audit_action_created", "action", "created_at"),
+    )
+
+    def __repr__(self) -> str:
+        return f"<AuditLog(action={self.action}, " f"resource_type={self.resource_type}, " f"outcome={self.outcome})>"
+
+
+# Standard audit actions
+class AuditAction:
+    """Standard audit action constants."""
+
+    # Authentication
+    LOGIN_SUCCESS = "login_success"
+    LOGIN_FAILED = "login_failed"
+    LOGOUT = "logout"
+    PASSWORD_CHANGED = "password_changed"  # nosec B105 - enum value, not a credential
+    PASSWORD_RESET_REQUESTED = "password_reset_requested"  # nosec B105
+    PASSWORD_RESET_COMPLETED = "password_reset_completed"  # nosec B105
+
+    # MFA
+    MFA_ENABLED = "mfa_enabled"
+    MFA_DISABLED = "mfa_disabled"
+    MFA_VERIFIED = "mfa_verified"
+    MFA_FAILED = "mfa_failed"
+    BACKUP_CODE_USED = "backup_code_used"
+
+    # SSO
+    SSO_LOGIN = "sso_login"
+    SSO_LINKED = "sso_linked"
+    SSO_UNLINKED = "sso_unlinked"
+
+    # Users
+    USER_CREATED = "user_created"
+    USER_UPDATED = "user_updated"
+    USER_DELETED = "user_deleted"
+    USER_ACTIVATED = "user_activated"
+    USER_DEACTIVATED = "user_deactivated"
+
+    # Teams
+    TEAM_CREATED = "team_created"
+    TEAM_UPDATED = "team_updated"
+    TEAM_DELETED = "team_deleted"
+    TEAM_MEMBER_ADDED = "team_member_added"
+    TEAM_MEMBER_REMOVED = "team_member_removed"
+    TEAM_MEMBER_ROLE_CHANGED = "team_member_role_changed"
+
+    # Organizations
+    ORG_CREATED = "org_created"
+    ORG_UPDATED = "org_updated"
+    ORG_DELETED = "org_deleted"
+
+    # Roles & Permissions
+    ROLE_CREATED = "role_created"
+    ROLE_UPDATED = "role_updated"
+    ROLE_DELETED = "role_deleted"
+    ROLE_ASSIGNED = "role_assigned"
+    ROLE_REVOKED = "role_revoked"
+
+    # API Keys
+    API_KEY_CREATED = "api_key_created"
+    API_KEY_REVOKED = "api_key_revoked"
+    API_KEY_USED = "api_key_used"
+
+    # Settings
+    SETTINGS_UPDATED = "settings_updated"
+
+    # Security
+    ACCOUNT_LOCKED = "account_locked"
+    ACCOUNT_UNLOCKED = "account_unlocked"
+    SUSPICIOUS_ACTIVITY = "suspicious_activity"
+
+    # RBAC (GH #6511)
+    PERMISSION_DENIED = "permission_denied"
+
+
+class AuditResourceType:
+    """Standard resource type constants."""
+
+    USER = "user"
+    TEAM = "team"
+    ORGANIZATION = "organization"
+    ROLE = "role"
+    PERMISSION = "permission"
+    API_KEY = "api_key"  # nosemgrep: autobot-hardcoded-secret-key
+    SSO_PROVIDER = "sso_provider"
+    SESSION = "session"
+    SETTINGS = "settings"
+    ENDPOINT = "endpoint"  # GH #6511: resource type for permission-deny audit events
